use poutine.queue in infer.enum and move nonlocalexit to poutine.util
