+++ pyro/poutine/__init__.py
@@ -2,7 +2,8 @@ from __future__ import absolute_import, division, print_function
 
 import functools
 
-from pyro import util
+import pyro.util
+from .util import NonlocalExit
 
 # poutines
 from .block_poutine import BlockPoutine
@@ -193,12 +194,10 @@ def queue(fn, queue, max_tries=None,
         max_tries = int(1e6)
 
     if extend_fn is None:
-        # XXX should be util.enum_extend
-        extend_fn = util.enum_extend
+        extend_fn = pyro.util.enum_extend
 
     if escape_fn is None:
-        # XXX should be util.discrete_escape
-        escape_fn = util.discrete_escape
+        escape_fn = pyro.util.discrete_escape
 
     if num_samples is None:
         num_samples = -1
@@ -214,9 +213,8 @@ def queue(fn, queue, max_tries=None,
                 ftr = trace(escape(replay(fn, next_trace),
                                    functools.partial(escape_fn, next_trace)))
                 return ftr(*args, **kwargs)
-            except util.NonlocalExit as site_container:
-                for frame in _PYRO_STACK:
-                    frame._reset()
+            except NonlocalExit as site_container:
+                site_container.reset_stack()
                 for tr in extend_fn(ftr.trace.copy(), site_container.site,
                                     num_samples=num_samples):
                     queue.put(tr)
