+++ pyro/poutine/escape_poutine.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from pyro.util import NonlocalExit
+from .util import NonlocalExit
 
 from .poutine import Messenger, Poutine
